fix: cast Buffer to Uint8Array for createDecipheriv
@types/node 20+ tightened CipherKey/ArrayBufferView typings; Buffer<ArrayBufferLike>
no longer satisfies them. Cast at call sites in byod-run-migrations and
byod-health-check to unblock @blah-chat/jobs DTS build.

+++ packages/jobs/src/trigger/byod-run-migrations.ts
@@ -165,8 +165,12 @@ async function defaultDecrypt(
     rawKey.length === 64
       ? Buffer.from(rawKey, "hex")
       : createHash("sha256").update(rawKey).digest();
-  const decipher = createDecipheriv("aes-256-gcm", key, Buffer.from(iv, "hex"));
-  decipher.setAuthTag(Buffer.from(authTag, "hex"));
+  const decipher = createDecipheriv(
+    "aes-256-gcm",
+    key as Uint8Array,
+    Buffer.from(iv, "hex") as Uint8Array,
+  );
+  decipher.setAuthTag(Buffer.from(authTag, "hex") as Uint8Array);
   let decrypted = decipher.update(encrypted, "hex", "utf8");
   decrypted += decipher.final("utf8");
   return decrypted;
